Moved original implementations of Router LAM and Update Generator to the examples folder
