+++ mainproject/cluster/src/main/resources/application.conf
@@ -1,8 +1,8 @@
 settings {
-	hostname = ${?HOSTNAME}
-	ip		   = ${?HOST_IP} // original was localhost
-	//hostName = "localhost"
-	//ip = "localhost"
+	//hostname = ${?HOSTNAME}
+	//ip		   = ${?HOST_IP} // original was localhost
+	hostName = "localhost"
+	ip = "localhost"
 	http 		 = 8080
 	port 		 = 1600
 	bport		 = 1600
