+++ mainproject/cluster/src/main/scala/com/raphtory/examples/random/actors/TestLAM.scala
@@ -1,10 +1,13 @@
-package com.raphtory.core.actors.analysismanager
+package com.raphtory.examples.random.actors
+
+import com.raphtory.core.actors.analysismanager.LiveAnalysisManager
 import com.raphtory.core.analysis._
+import com.raphtory.examples.random.analysis.TestAnalyser
 
 class TestLAM extends LiveAnalysisManager {
   override protected def defineMaxSteps(): Int = 10
 
-  override protected def generateAnalyzer: Analyser = new TestAnalyser2()
+  override protected def generateAnalyzer: Analyser = new TestAnalyser()
 
   override protected def processResults(result: Any): Unit = println(result)
