+++ mainproject/cluster/src/main/scala/com/raphtory/tests/SingleNodeTest.scala
@@ -11,16 +11,20 @@ object SingleNodeTest extends App {
 
   val partitionNumber = 1
   val minimumRouters = 1
-  val routerClassName = "com.raphtory.examples.bitcoin.actors.BitcoinRouter"
-  val LamClassName = "com.raphtory.core.actors.analysismanager.TestLAM"
-  val UpdaterName = "com.raphtory.examples.bitcoin.actors.BitcoinSpout"
+  //val routerClassName = "com.raphtory.examples.bitcoin.actors.BitcoinRouter"
+  //val LamClassName = "com.raphtory.examples.random.actors.TestLAM"
+  //val UpdaterName = "com.raphtory.examples.bitcoin.actors.BitcoinSpout"
+  val routerClassName = "com.raphtory.examples.random.actors.RaphtoryRouter"
+  val LamClassName = "com.raphtory.examples.random.actors.TestLAM"
+  val UpdaterName = "com.raphtory.examples.random.actors.UpdateGen"
+
   val system = ActorSystem("Single-Node-test")
 
   system.actorOf(Props(new WatchDog(partitionNumber,minimumRouters)), "WatchDog")
   system.actorOf(Props(RaphtoryReplicator("Router", routerClassName)), s"Routers")
   system.actorOf(Props(RaphtoryReplicator("Partition Manager")), s"PartitionManager")
   system.actorOf(Props(Class.forName(UpdaterName)), "UpdateGen")
-  //system.actorOf(Props(Class.forName(LamClassName)), s"LiveAnalysisManager_$LamClassName")
+  system.actorOf(Props(Class.forName(LamClassName)), s"LiveAnalysisManager_$LamClassName")
 
 }
